Refactor code to split across different files
